+++ dotnet/src/Downstairs.Infrastructure/Persistence/Models/OrderFixedPriceRow.cs
@@ -10,7 +10,7 @@ public partial class OrderFixedPriceRow
 
     public string? Description { get; set; }
 
-    public uint Quantity { get; set; }
+    public long Quantity { get; set; }
 
     public decimal Price { get; set; }
